requirements.txt
fastapi
uvicorn[standard]
pandas
torch
torchvision
sentence-transformers
python-multipart
Pillow